+++ internal/cmd/local/local_install.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"github.com/airbytehq/abctl/internal/cmd/local/docker"
 	k8s2 "github.com/airbytehq/abctl/internal/cmd/local/k8s"
-	"github.com/airbytehq/abctl/internal/local"
+	"github.com/airbytehq/abctl/internal/cmd/local/local"
 	"github.com/airbytehq/abctl/internal/telemetry"
 	"github.com/pterm/pterm"
 	"github.com/spf13/cobra"
